move internal contexts from CurvePrimitive.ts; reorder imports using Code extension "TypeScript Import Sorter" (Ctrl-Alt-O)
