+++ core/geometry/src/curve/CurvePrimitive.ts
@@ -6,35 +6,36 @@
  * @module Curve
  */
 import { assert } from "@itwin/core-bentley";
-import type { InterpolationCurve3d } from "../bspline/InterpolationCurve3d";
 import { Clipper } from "../clipping/ClipUtils";
 import { StrokeCountMap } from "../curve/Query/StrokeCountMap";
 import { AxisOrder, Geometry, PlaneAltitudeEvaluator } from "../Geometry";
+import { IStrokeHandler } from "../geometry3d/GeometryHandler";
+import { Matrix3d } from "../geometry3d/Matrix3d";
+import { Plane3dByOriginAndUnitNormal } from "../geometry3d/Plane3dByOriginAndUnitNormal";
+import { Plane3dByOriginAndVectors } from "../geometry3d/Plane3dByOriginAndVectors";
+import { Point3d } from "../geometry3d/Point3dVector3d";
+import { Range3d } from "../geometry3d/Range";
+import { Ray3d } from "../geometry3d/Ray3d";
+import { Transform } from "../geometry3d/Transform";
+import { VariantCurveExtendParameter } from "./CurveExtendMode";
+import { CurveIntervalRole, CurveLocationDetail, CurveSearchStatus } from "./CurveLocationDetail";
+import { GeometryQuery } from "./GeometryQuery";
+import { AppendPlaneIntersectionStrokeHandler } from "./internalContexts/AppendPlaneIntersectionStrokeHandler";
+import { ClosestPointStrokeHandler } from "./internalContexts/ClosestPointStrokeHandler";
+import { CurveLengthContext } from "./internalContexts/CurveLengthContext";
+import { LineString3d } from "./LineString3d";
+import { StrokeOptions } from "./StrokeOptions";
+
+import type { InterpolationCurve3d } from "../bspline/InterpolationCurve3d";
 import type { AkimaCurve3d } from "../bspline/AkimaCurve3d";
 import type { BSplineCurve3d } from "../bspline/BSplineCurve";
 import type { BezierCurve3d } from "../bspline/BezierCurve3d";
 import type { CurveChainWithDistanceIndex } from "./CurveChainWithDistanceIndex";
 import type { DirectSpiral3d } from "./spiral/DirectSpiral3d";
 import type { IntegratedSpiral3d } from "./spiral/IntegratedSpiral3d";
-import { IStrokeHandler } from "../geometry3d/GeometryHandler";
-import { Matrix3d } from "../geometry3d/Matrix3d";
-import { Plane3dByOriginAndUnitNormal } from "../geometry3d/Plane3dByOriginAndUnitNormal";
-import { Plane3dByOriginAndVectors } from "../geometry3d/Plane3dByOriginAndVectors";
-import { Point3d, Vector3d } from "../geometry3d/Point3dVector3d";
-import { Ray3d } from "../geometry3d/Ray3d";
-import { Transform } from "../geometry3d/Transform";
-import { Order2Bezier } from "../numerics/BezierPolynomials";
-import { Newton1dUnboundedApproximateDerivative, NewtonEvaluatorRtoR } from "../numerics/Newton";
-import { GaussMapper } from "../numerics/Quadrature";
 import type { Arc3d } from "./Arc3d";
-import { CurveExtendOptions, VariantCurveExtendParameter } from "./CurveExtendMode";
-import { CurveIntervalRole, CurveLocationDetail, CurveSearchStatus } from "./CurveLocationDetail";
-import { GeometryQuery } from "./GeometryQuery";
 import type { LineSegment3d } from "./LineSegment3d";
-import { LineString3d } from "./LineString3d";
-import { StrokeOptions } from "./StrokeOptions";
 import type { OffsetOptions } from "./internalContexts/PolygonOffsetContext";
-import { Range1d, Range3d } from "../geometry3d/Range";
 
 /** Describes the concrete type of a [[CurvePrimitive]]. Each type name maps to a specific subclass and can be used for type-switching in conditional statements.
  *  - "arc" => [[Arc3d]]
@@ -740,369 +741,4 @@ export abstract class CurvePrimitive extends GeometryQuery {
    * @param offsetDistanceOrOptions offset distance (positive to left of the instance curve), or options object
    */
   public abstract constructOffsetXY(offsetDistanceOrOptions: number | OffsetOptions): CurvePrimitive | CurvePrimitive[] | undefined;
-}
-
-/** Intermediate class for managing the parentCurve announcements from an IStrokeHandler */
-abstract class NewtonRotRStrokeHandler extends NewtonEvaluatorRtoR {
-  protected _parentCurvePrimitive: CurvePrimitive | undefined;
-  constructor() {
-    super();
-    this._parentCurvePrimitive = undefined;
-  }
-  /** retain the parentCurvePrimitive.
-   * * Calling this method tells the handler that the parent curve is to be used for detail searches.
-   * * Example: Transition spiral search is based on linestring first, then the exact spiral.
-   * * Example: CurveChainWithDistanceIndex does NOT do this announcement -- the constituents act independently.
-   */
-  public startParentCurvePrimitive(curve: CurvePrimitive | undefined) { this._parentCurvePrimitive = curve; }
-  /** Forget the parentCurvePrimitive */
-  public endParentCurvePrimitive(_curve: CurvePrimitive | undefined) { this._parentCurvePrimitive = undefined; }
-}
-
-class AppendPlaneIntersectionStrokeHandler extends NewtonRotRStrokeHandler implements IStrokeHandler {
-  private _curve: CurvePrimitive | undefined;
-  private _plane: PlaneAltitudeEvaluator;
-  private _intersections: CurveLocationDetail[];
-  private _fractionA: number = 0;
-  private _functionA: number = 0;
-  // private derivativeA: number;   <---- Not currently used
-  private _functionB: number = 0;
-  private _fractionB: number = 0;
-  private _derivativeB: number = 0;
-  private _numThisCurve: number = 0;
-  // scratch vars for use within methods.
-  private _ray: Ray3d;
-  private _newtonSolver: Newton1dUnboundedApproximateDerivative;
-
-  // Return the first defined curve among: this.parentCurvePrimitive, this.curve;
-  public effectiveCurve(): CurvePrimitive | undefined {
-    if (this._parentCurvePrimitive)
-      return this._parentCurvePrimitive;
-    return this._curve;
-  }
-  public get getDerivativeB() { return this._derivativeB; }    // <--- DerivativeB is not currently used anywhere. Provided getter to suppress lint error
-
-  public constructor(plane: PlaneAltitudeEvaluator, intersections: CurveLocationDetail[]) {
-    super();
-    this._plane = plane;
-    this._intersections = intersections;
-    this.startCurvePrimitive(undefined);
-    this._ray = Ray3d.createZero();
-    this._newtonSolver = new Newton1dUnboundedApproximateDerivative(this);
-  }
-  public startCurvePrimitive(curve: CurvePrimitive | undefined) {
-    this._curve = curve;
-    this._fractionA = 0.0;
-    this._numThisCurve = 0;
-    this._functionA = 0.0;
-    // this.derivativeA = 0.0;
-  }
-  public endCurvePrimitive() { }
-  public announceIntervalForUniformStepStrokes(
-    cp: CurvePrimitive,
-    numStrokes: number,
-    fraction0: number,
-    fraction1: number): void {
-    this.startCurvePrimitive(cp);
-    if (numStrokes < 1) numStrokes = 1;
-    const df = 1.0 / numStrokes;
-    for (let i = 0; i <= numStrokes; i++) {
-      const fraction = Geometry.interpolate(fraction0, i * df, fraction1);
-      cp.fractionToPointAndDerivative(fraction, this._ray);
-      this.announcePointTangent(this._ray.origin, fraction, this._ray.direction);
-    }
-  }
-  public announceSegmentInterval(
-    _cp: CurvePrimitive,
-    point0: Point3d,
-    point1: Point3d,
-    _numStrokes: number,
-    fraction0: number,
-    fraction1: number): void {
-    const h0 = this._plane.altitude(point0);
-    const h1 = this._plane.altitude(point1);
-    if (h0 * h1 > 0.0)
-      return;
-    const fraction01 = Order2Bezier.solveCoffs(h0, h1);
-    // let numIntersection = 0;
-    if (fraction01 !== undefined) {
-      // numIntersection++;
-      const fraction = Geometry.interpolate(fraction0, fraction01, fraction1);
-      this._newtonSolver.setX(fraction);
-      if (this._newtonSolver.runIterations()) {
-        this.announceSolutionFraction(this._newtonSolver.getX());
-      }
-      // this.intersections.push(CurveLocationDetail.createCurveFractionPoint(cp, fraction, cp.fractionToPoint(fraction)));
-    }
-  }
-  private announceSolutionFraction(fraction: number) {
-    const curve = this.effectiveCurve();
-    if (curve) {
-      this._ray = curve.fractionToPointAndDerivative(fraction, this._ray);
-      this._intersections.push(CurveLocationDetail.createCurveFractionPoint(curve, fraction, this._ray.origin));
-    }
-  }
-  public evaluate(fraction: number): boolean {
-    const curve = this.effectiveCurve();
-    if (!curve)
-      return false;
-    this.currentF = this._plane.altitude(curve.fractionToPoint(fraction));
-    return true;
-  }
-  /**
-   * * ASSUME both the "A" and "B"  evaluations (fraction, function, and derivative) are known.
-   * * If function value changed sign between, interpolate an approximate root and improve it with
-   *     the newton solver.
-   */
-  private searchInterval() {
-    if (this._functionA * this._functionB > 0) return;
-    if (this._functionA === 0) this.announceSolutionFraction(this._fractionA);
-    if (this._functionB === 0) this.announceSolutionFraction(this._fractionB);
-    if (this._functionA * this._functionB < 0) {
-      const fraction = Geometry.inverseInterpolate(this._fractionA, this._functionA, this._fractionB, this._functionB);
-      if (fraction) {
-        this._newtonSolver.setX(fraction);
-        if (this._newtonSolver.runIterations())
-          this.announceSolutionFraction(this._newtonSolver.getX());
-      }
-    }
-  }
-  /** Evaluate and save _functionB, _derivativeB, and _fractionB. */
-  private evaluateB(xyz: Point3d, fraction: number, tangent: Vector3d) {
-    this._functionB = this._plane.altitude(xyz);
-    this._derivativeB = this._plane.velocity(tangent);
-    this._fractionB = fraction;
-  }
-  /**
-   * Announce point and tangent for evaluations.
-   * * The function evaluation is saved as the "B" function point.
-   * * The function point count is incremented
-   * * If function point count is greater than 1, the current interval is searched.
-   * * The just-evaluated point ("B") is saved as the "old" ("A") evaluation point.
-   * @param xyz
-   * @param fraction
-   * @param tangent
-   */
-  public announcePointTangent(xyz: Point3d, fraction: number, tangent: Vector3d): void {
-    this.evaluateB(xyz, fraction, tangent);
-    if (this._numThisCurve++ > 0) this.searchInterval();
-    this._functionA = this._functionB;
-    this._fractionA = this._fractionB;
-  }
-}
-
-export class CurveLengthContext implements IStrokeHandler {
-  private _curve: CurvePrimitive | undefined;
-  private _summedLength: number;
-  private _ray: Ray3d;
-  private _fraction0: number;
-  private _fraction1: number;
-  private _gaussMapper: GaussMapper;
-
-  private tangentMagnitude(fraction: number): number {
-    this._ray = (this._curve as CurvePrimitive).fractionToPointAndDerivative(fraction, this._ray);
-    return this._ray.direction.magnitude();
-  }
-
-  /** Return the fraction0 installed at construction time. */
-  public get getFraction0(): number {
-    return this._fraction0;
-  }
-
-  /** Return the fraction1 installed at construction time. */
-  public get getFraction1(): number {
-    return this._fraction1;
-  }
-
-  public getSum() { return this._summedLength; }
-
-  public constructor(fraction0: number = 0.0, fraction1: number = 1.0, numGaussPoints: number = 5) {
-    this.startCurvePrimitive(undefined);
-    this._summedLength = 0.0;
-    this._ray = Ray3d.createZero();
-    if (fraction0 < fraction1) {
-      this._fraction0 = fraction0;
-      this._fraction1 = fraction1;
-    } else {
-      this._fraction0 = fraction1;
-      this._fraction1 = fraction0;
-    }
-    this._gaussMapper = new GaussMapper(numGaussPoints);
-  }
-  public startCurvePrimitive(curve: CurvePrimitive | undefined) {
-    this._curve = curve;
-  }
-  public startParentCurvePrimitive(_curve: CurvePrimitive) { }
-  public endParentCurvePrimitive(_curve: CurvePrimitive) { }
-
-  public endCurvePrimitive() { }
-  public announceIntervalForUniformStepStrokes(cp: CurvePrimitive, numStrokes: number, fraction0: number, fraction1: number): void {
-    const range = Range1d.createXX(fraction0, fraction1);
-    range.intersectRangeXXInPlace(this._fraction0, this._fraction1);
-    if (!range.isNull) {
-      this.startCurvePrimitive(cp);
-      if (numStrokes < 1) numStrokes = 1;
-      const df = 1.0 / numStrokes;
-      for (let i = 1; i <= numStrokes; i++) {
-        const fractionA = range.fractionToPoint((i - 1) * df);
-        const fractionB = i === numStrokes ? range.high : range.fractionToPoint(i * df);
-        const numGauss = this._gaussMapper.mapXAndW(fractionA, fractionB);
-        for (let k = 0; k < numGauss; k++) {
-          this._summedLength += this._gaussMapper.gaussW[k] * this.tangentMagnitude(this._gaussMapper.gaussX[k]);
-        }
-      }
-    }
-  }
-  public announceSegmentInterval(_cp: CurvePrimitive, point0: Point3d, point1: Point3d, _numStrokes: number, fraction0: number, fraction1: number): void {
-    const segmentLength = point0.distance(point1);
-    if (this._fraction0 <= fraction0 && fraction1 <= this._fraction1)
-      this._summedLength += segmentLength;
-    else {
-      const range = Range1d.createXX(fraction0, fraction1);
-      range.intersectRangeXXInPlace(this._fraction0, this._fraction1);
-      if (!range.isNull)
-        this._summedLength += segmentLength * range.length() / (fraction1 - fraction0);
-    }
-  }
-
-  public announcePointTangent(_xyz: Point3d, _fraction: number, _tangent: Vector3d): void {
-    // uh oh -- need to retain point for next interval
-  }
-}
-// context for searching for closest point .. .
-class ClosestPointStrokeHandler extends NewtonRotRStrokeHandler implements IStrokeHandler {
-  private _curve: CurvePrimitive | undefined;
-  private _closestPoint: CurveLocationDetail | undefined;
-  private _spacePoint: Point3d;
-  private _extend: VariantCurveExtendParameter;
-  private _fractionA: number = 0;
-  private _functionA: number = 0;
-  private _functionB: number = 0;
-  private _fractionB: number = 0;
-  private _numThisCurve: number = 0;
-  // scratch vars for use within methods.
-  private _workPoint: Point3d;
-  private _workRay: Ray3d;
-  private _newtonSolver: Newton1dUnboundedApproximateDerivative;
-
-  public constructor(spacePoint: Point3d, extend: VariantCurveExtendParameter) {
-    super();
-    this._spacePoint = spacePoint;
-    this._workPoint = Point3d.create();
-    this._workRay = Ray3d.createZero();
-    this._closestPoint = undefined;
-    this._extend = extend;
-    this.startCurvePrimitive(undefined);
-    this._newtonSolver = new Newton1dUnboundedApproximateDerivative(this);
-  }
-
-  public claimResult(): CurveLocationDetail | undefined {
-    if (this._closestPoint) {
-      this._newtonSolver.setX(this._closestPoint.fraction);
-      this._curve = this._closestPoint.curve;
-      if (this._newtonSolver.runIterations()) {
-        let fraction = this._newtonSolver.getX();
-        fraction = CurveExtendOptions.correctFraction(this._extend, fraction);
-        this.announceSolutionFraction(fraction);
-      }
-    }
-    return this._closestPoint;
-  }
-  public needPrimaryGeometryForStrokes() { return true; }
-  public startCurvePrimitive(curve: CurvePrimitive | undefined) {
-    this._curve = curve;
-    this._fractionA = 0.0;
-    this._numThisCurve = 0;
-    this._functionA = 0.0;
-  }
-  public endCurvePrimitive() { }
-  public announceIntervalForUniformStepStrokes(
-    cp: CurvePrimitive,
-    numStrokes: number,
-    fraction0: number,
-    fraction1: number): void {
-    this.startCurvePrimitive(cp);
-    this.announceSolutionFraction(0.0); // test start point as closest
-    this.announceSolutionFraction(1.0); // test end point as closest
-    if (numStrokes < 1) numStrokes = 1;
-    const df = 1.0 / numStrokes;
-    for (let i = 0; i <= numStrokes; i++) {
-      const fraction = Geometry.interpolate(fraction0, i * df, fraction1);
-      cp.fractionToPointAndDerivative(fraction, this._workRay);
-      this.announceRay(fraction, this._workRay);
-    }
-  }
-
-  private announceCandidate(cp: CurvePrimitive, fraction: number, point: Point3d) {
-    const distance = this._spacePoint.distance(point);
-    if (this._closestPoint && distance > this._closestPoint.a)
-      return;
-    this._closestPoint = CurveLocationDetail.createCurveFractionPoint(cp, fraction, point, this._closestPoint);
-    this._closestPoint.a = distance;
-    if (this._parentCurvePrimitive !== undefined)
-      this._closestPoint.curve = this._parentCurvePrimitive;
-  }
-  public announceSegmentInterval(
-    cp: CurvePrimitive,
-    point0: Point3d,
-    point1: Point3d,
-    _numStrokes: number,
-    fraction0: number,
-    fraction1: number): void {
-    let localFraction = this._spacePoint.fractionOfProjectionToLine(point0, point1, 0.0);
-    // only consider extending the segment if the immediate caller says we are at endpoints ...
-    if (!this._extend)
-      localFraction = Geometry.clampToStartEnd(localFraction, 0.0, 1.0);
-    else {
-      if (fraction0 !== 0.0)
-        localFraction = Math.max(localFraction, 0.0);
-      if (fraction1 !== 1.0)
-        localFraction = Math.min(localFraction, 1.0);
-    }
-    this._workPoint = point0.interpolate(localFraction, point1);
-    const globalFraction = Geometry.interpolate(fraction0, localFraction, fraction1);
-    this.announceCandidate(cp, globalFraction, this._workPoint);
-  }
-  private searchInterval() {
-    if (this._functionA * this._functionB > 0) return;
-    if (this._functionA === 0) this.announceSolutionFraction(this._fractionA);
-    if (this._functionB === 0) this.announceSolutionFraction(this._fractionB);
-    if (this._functionA * this._functionB < 0) {
-      const fraction = Geometry.inverseInterpolate(this._fractionA, this._functionA, this._fractionB, this._functionB);
-      if (fraction) {
-        this._newtonSolver.setX(fraction);
-        if (this._newtonSolver.runIterations())
-          this.announceSolutionFraction(this._newtonSolver.getX());
-      }
-    }
-  }
-  private evaluateB(fractionB: number, dataB: Ray3d) {
-    this._functionB = dataB.dotProductToPoint(this._spacePoint);
-    this._fractionB = fractionB;
-  }
-  private announceSolutionFraction(fraction: number) {
-    if (this._curve)
-      this.announceCandidate(this._curve, fraction, this._curve.fractionToPoint(fraction));
-  }
-  public evaluate(fraction: number): boolean {
-    let curve = this._curve;
-    if (this._parentCurvePrimitive)
-      curve = this._parentCurvePrimitive;
-    if (curve) {
-      this._workRay = curve.fractionToPointAndDerivative(fraction, this._workRay);
-      this.currentF = this._workRay.dotProductToPoint(this._spacePoint);
-      return true;
-    }
-    return false;
-  }
-  public announceRay(fraction: number, data: Ray3d): void {
-    this.evaluateB(fraction, data);
-    if (this._numThisCurve++ > 0) this.searchInterval();
-    this._functionA = this._functionB;
-    this._fractionA = this._fractionB;
-  }
-  public announcePointTangent(point: Point3d, fraction: number, tangent: Vector3d) {
-    this._workRay.set(point, tangent);
-    this.announceRay(fraction, this._workRay);
-  }
 }
